requirements.txt
flask
python-dotenv
pymongo[srv]==3.12
ObjectId